DirectedAcyclicGraphs

DirectedAcyclicGraphs.DAG — Type

A node in a directed acyclic graph (of which it is the root)

DirectedAcyclicGraphs.Inner — Type

The trait of inner nodes (nodes that have children)

DirectedAcyclicGraphs.Leaf — Type

The trait of leaf nodes (nodes without children)

DirectedAcyclicGraphs.NodeType — Type

A trait hierarchy denoting types of DAG nodes NodeType defines an orthogonal type hierarchy of node types, so we can dispatch on node type regardless of the graph type. See @ref{https://docs.julialang.org/en/v1/manual/methods/#Trait-based-dispatch-1}

DirectedAcyclicGraphs.NodeType — Method

Get the node type trait of the given Node

DirectedAcyclicGraphs.Tree — Type

A node in a tree (of which it is the root)

Base.filter — Method

Retrieve list of nodes in DAG matching predicate p

Base.foreach — Function

Apply a function to each node in a graph, bottom up

Base.in — Function

Is the node contained in the DAG?

Base.parent — Method

Get the parent of a given tree node (or nothing if the node is root)

DirectedAcyclicGraphs.children — Function

Get the children of a given inner node

DirectedAcyclicGraphs.depth — Method

Compute the length of the path from a tree node to the leaf following the branching function

DirectedAcyclicGraphs.find_inode — Method

Find a binary inner node that has left in its left subtree and right in its right subtree. Supports nothing as a catch-all for either left or right. Returns nothing if no such node exists.

DirectedAcyclicGraphs.find_leaf — Method

Find the leaf in the tree by follwing the branching function

DirectedAcyclicGraphs.foldup — Method

foldup(node::DAG, 
    f_leaf::Function, 
    f_inner::Function, 
    ::Type{T})::T where {T}

Compute a function bottom-up on the graph. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner as a function on the children.

DirectedAcyclicGraphs.foldup_aggregate — Method

Compute a function bottom-up on the circuit. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner in aggregate as a vector from the children.

DirectedAcyclicGraphs.foreach_down — Method

Apply a function to each node in a graph, top down

DirectedAcyclicGraphs.groupby — Method

Group the elements of list by their values according to function f

DirectedAcyclicGraphs.has_children — Method

Does the DAG node have children?

DirectedAcyclicGraphs.has_parent — Method

Does the node have a parent?

DirectedAcyclicGraphs.innernode_stats — Method

innernode_stats(c::DAG)

Give count of types and fan-ins of inner nodes in the graph

DirectedAcyclicGraphs.innernodes — Method

Get the list of inner nodes in a given graph

DirectedAcyclicGraphs.isequal_local — Method

Is one node equal to another locally, ignoring children?

DirectedAcyclicGraphs.isinner — Method

Is the DAG node an inner node?

DirectedAcyclicGraphs.isleaf — Method

Is the DAG node a leaf node?

DirectedAcyclicGraphs.isroot — Method

Is the node the root of its tree?

DirectedAcyclicGraphs.label_nodes — Method

Assign an integer label to each circuit node, bottom up, starting at 1

DirectedAcyclicGraphs.lca — Method

Find the least common ancestor. Assumes the Tree has access to a parent. A given descends_from function is required to quickly check whether a node is an ancestor.

DirectedAcyclicGraphs.leaf_stats — Method

leaf_stats(c::DAG)

Give count of types of leaf nodes in the graph

DirectedAcyclicGraphs.leafnodes — Method

Get the list of leaf nodes in a given graph

DirectedAcyclicGraphs.left_most_descendent — Method

left_most_descendent(root::DAG)::DAG

Return the left-most descendent.

DirectedAcyclicGraphs.linearize — Method

Order the DAG's nodes bottom-up in a list (with optional element type)

DirectedAcyclicGraphs.map_values — Method

Map the values in the dictionary, retaining the same keys

DirectedAcyclicGraphs.node_stats — Method

node_stats(c::DAG)

Give count of types and fan-ins of all nodes in the graph

DirectedAcyclicGraphs.num_children — Method

Get the number of children of a given inner DAG node

DirectedAcyclicGraphs.num_edges — Function

Number of edges in the DAG

DirectedAcyclicGraphs.num_innernodes — Method

Count the number of inner nodes in a given graph

DirectedAcyclicGraphs.num_leafnodes — Method

Count the number of leaf nodes in a given graph

DirectedAcyclicGraphs.num_nodes — Function

num_nodes(node::DAG)

Count the number of nodes in the DAG

DirectedAcyclicGraphs.num_parents — Method

num_parents(root::DAG)

Count the number of parents for each node in DAG under root

DirectedAcyclicGraphs.parent_stats — Method

parent_stats(c::DAG)

Give number of nodes grouped by (type, parent_count)

DirectedAcyclicGraphs.print_tree — Function

Print the given tree

DirectedAcyclicGraphs.right_most_descendent — Method

right_most_descendent(root::DAG)::DAG

Return the right-most descendent.

DirectedAcyclicGraphs.root — Method

Get the root of the given tree node

DirectedAcyclicGraphs.tree_num_edges — Function

tree_num_edges(node::DAG)::BigInt

Compute the number of edges in the tree-unfolding of the DAG.

DirectedAcyclicGraphs.tree_num_nodes — Function

tree_num_nodes(node::DAG)::BigInt

Compute the number of nodes in of a tree-unfolding of the DAG.